Move termination into the node directory
